Software/Testing/Publishers/AirControl/requirements.txt
# Arquivo de dependências do projeto
# Para instalar, execute: pip install -r requirements.txt

paho-mqtt==2.1.0